backend/fastapi/requirements.txt
# FastAPI Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.25.2
ta-lib==0.4.28

# Async & HTTP
aiohttp==3.9.1
httpx==0.25.2
websockets==12.0

# Caching
redis==5.0.1
hiredis==2.2.3

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
click==8.1.7

# Financial & Trading
yfinance==0.2.28
ccxt==4.1.64

# Image Processing (for charts)
pillow==10.1.0
matplotlib==3.8.2

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Task Queue (optional)
celery==5.3.4
flower==2.0.1

# API Documentation
fastapi-users==0.12.0
fastapi-pagination==0.12.13

# Development Server
watchdog==3.0.0